refactor(http): move unstable APIs (header, method, signed-cookie)
