+++ http/deno.json
@@ -6,12 +6,12 @@
     "./cookie": "./cookie.ts",
     "./etag": "./etag.ts",
     "./file-server": "./file_server.ts",
-    "./header": "./header.ts",
-    "./method": "./method.ts",
+    "./unstable-header": "./unstable_header.ts",
+    "./unstable-method": "./unstable_method.ts",
     "./negotiation": "./negotiation.ts",
     "./server-sent-event-stream": "./server_sent_event_stream.ts",
     "./status": "./status.ts",
-    "./signed-cookie": "./signed_cookie.ts",
+    "./unstable-signed-cookie": "./unstable_signed_cookie.ts",
     "./user-agent": "./user_agent.ts",
     "./route": "./route.ts"
   }
