+++ http/mod.ts
@@ -96,10 +96,7 @@ export * from "./cookie.ts";
 export * from "./etag.ts";
 export * from "./status.ts";
 export * from "./negotiation.ts";
-export * from "./signed_cookie.ts";
 export * from "./server_sent_event_stream.ts";
 export * from "./user_agent.ts";
 export * from "./file_server.ts";
 export * from "./route.ts";
-export * from "./header.ts";
-export * from "./method.ts";
